+++ client/src/views/major/MajorList.vue
@@ -16,7 +16,6 @@
         :data="list"
         stripe
         row-key="id"
-        :default-sort="{ prop: 'name', order: 'ascending' }"
       >
         <template #empty>
           <el-empty description="暂无专业数据，请点击右上角新增" />