+++ client/src/views/textbook/TextbookList.vue
@@ -56,7 +56,6 @@
         :data="filteredlist"
         stripe
         row-key="id"
-        :default-sort="{ prop: 'title', order: 'ascending' }"
         @selection-change="handleSelectionChange"
       >
         <template #empty>